% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateHaze.R
\name{estimateHaze}
\alias{estimateHaze}
\title{Estimate image haze for dark object subtraction procedures}
\usage{
estimateHaze(x, hazeBands, darkProp = 0.02, plot = FALSE,
  returnTables = FALSE)
}
\arguments{
\item{x}{raster object or a previous result from \code{estimateSHV(x , returnTables = TRUE} from which to estimate haze}

\item{hazeBands}{character. Band or bandname from which to estimate SHV (optional if x contains only one layer)}

\item{darkProp}{proportion of pixels estimated to be dark}

\item{plot}{display histograms and haze values}

\item{returnTables}{return the frequency table per layer. Only takes effect if x is a Raster* object. If x is a result of estimateSHV tables will always be returned.}
}
\description{
Estimate image haze for dark object subtraction procedures
}

