% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Extract bandwise information from ImageMetaData}
\usage{
getMeta(img, metaData, what)
}
\arguments{
\item{img}{Raster* or character vector with band names.}

\item{metaData}{ImageMetaData or path to meta data file.}

\item{what}{Character. Slot to extract.}
}
\description{
This is an accessor function to quickly access information stored in ImageMetaData, e.g. scale factor per band.
Intended for use with imagery which was imported using stackMeta.
}

