% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNumeric}
\alias{getNumeric}
\title{Extract numbers from strings}
\usage{
getNumeric(x, returnNumeric = TRUE)
}
\arguments{
\item{x}{string or vector of strings}

\item{returnNumeric}{logical. should results be formatted \code{as.numeric}? If so, "05" will be converted to 5. Set returnNumeric to \code{FALSE} to keep preceeding zeros.}
}
\description{
Extract numbers from strings
}
\note{
decimal numbers will be returned as two separate numbers
}

