% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{LANDSAT.db}
\alias{LANDSAT.db}
\title{Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php}
\format{\preformatted{List of 3
 $ LANDSAT5:List of 1
  ..$ TM:'data.frame':	7 obs. of  11 variables:
  .. ..$ band      : Factor w/ 7 levels "B1","B2","B3",..: 1 2 3 4 5 6 7
  .. ..$ bandtype  : Factor w/ 2 levels "REF","TIR": 1 1 1 1 1 2 1
  .. ..$ centerWavl: num [1:7] 0.485 0.569 0.66 0.84 1.676 ...
  .. ..$ spatRes1  : num [1:7] 30 30 30 30 30 30 30
  .. ..$ spatRes2  : num [1:7] 30 30 30 30 30 60 30
  .. ..$ esun      : num [1:7] 1983 1796 1536 1031 220 ...
  .. ..$ veryClear : num [1:7] 18.073 9.54 5.27 2.009 0.127 ...
  .. ..$ clear     : num [1:7] 4.251 3.089 2.296 1.417 0.356 ...
  .. ..$ moderate  : num [1:7] 2.062 1.757 1.515 1.19 0.597 ...
  .. ..$ hazy      : num [1:7] 1.66 1.484 1.338 1.13 0.697 ...
  .. ..$ veryHazy  : num [1:7] 1.436 1.326 1.231 1.091 0.772 ...
 $ LANDSAT7:List of 1
  ..$ ETM:'data.frame':	8 obs. of  11 variables:
  .. ..$ band      : Factor w/ 8 levels "B1","B2","B3",..: 1 2 3 4 5 6 7 8
  .. ..$ bandtype  : Factor w/ 3 levels "PAN","REF","TIR": 2 2 2 2 2 3 2 1
  .. ..$ spatRes1  : num [1:8] 30 30 30 30 30 30 30 15
  .. ..$ spatRes2  : num [1:8] 30 30 30 30 30 60 30 15
  .. ..$ centerWavl: num [1:8] 0.483 0.56 0.662 0.825 1.648 ...
  .. ..$ esun      : num [1:8] 1997 1812 1533 1039 231 ...
  .. ..$ veryClear : num [1:8] 18.374 10.168 5.207 2.159 0.136 ...
  .. ..$ clear     : num [1:8] 4.287 3.189 2.282 1.469 0.368 ...
  .. ..$ moderate  : num [1:8] 2.07 1.786 1.511 1.212 0.607 ...
  .. ..$ hazy      : num [1:8] 1.664 1.501 1.335 1.144 0.705 ...
  .. ..$ veryHazy  : num [1:8] 1.439 1.336 1.229 1.101 0.779 ...
 $ LANDSAT8:List of 1
  ..$ OLI_TIRS:'data.frame':	12 obs. of  11 variables:
  .. ..$ band      : Factor w/ 12 levels "B1","B10","B11",..: 1 4 5 6 7 8 9 10 11 2 ...
  .. ..$ bandtype  : Factor w/ 4 levels "PAN","QA","REF",..: 3 3 3 3 3 3 3 1 3 4 ...
  .. ..$ spatRes1  : num [1:12] 30 30 30 30 30 30 30 15 30 30 ...
  .. ..$ spatRes2  : num [1:12] 30 30 30 30 30 30 30 15 30 30 ...
  .. ..$ centerWavl: num [1:12] 0.44 0.48 0.56 0.655 0.865 1.61 2.2 0.59 1.37 10.6 ...
  .. ..$ esun      : num [1:12] NA 2067 1893 1603 973 ...
  .. ..$ veryClear : num [1:12] 26.68 18.84 10.17 5.43 1.79 ...
  .. ..$ clear     : num [1:12] 5.17 4.34 3.19 2.33 1.34 ...
  .. ..$ moderate  : num [1:12] 2.27 2.08 1.79 1.53 1.16 ...
  .. ..$ hazy      : num [1:12] 1.78 1.67 1.5 1.34 1.11 ...
  .. ..$ veryHazy  : num [1:12] 1.51 1.44 1.34 1.24 1.08 ...
}}
\usage{
LANDSAT.db
}
\description{
Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php
}
\keyword{datasets}

