% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox-package.R
\docType{data}
\name{lsat}
\alias{lsat}
\title{Landsat 5TM Example Data}
\usage{
lsat
}
\description{
Subset of Landsat 5 TM Scene: LT52240631988227CUB02
Contains all seven bands in DN format.
}
\examples{
ggRGB(lsat, stretch = "sqrt")
}
\keyword{datasets}
