% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesma.R
\name{mesma}
\alias{mesma}
\title{Multiple Endmember Spectral Mixture Analysis (Spectral Unmixing)}
\usage{
mesma(img, em, method = "NNLS", iterate = 400, tolerance = 1e-08, ...,
  verbose)
}
\arguments{
\item{img}{RasterBrick or RasterStack. Remote sensing imagery (usually hyperspectral).}

\item{em}{Matrix or data.frame with spectral endmembers. Rows represent a single endmember of a class, columns represent the spectral bands (i.e. columns correspond to number of bands in \code{img}). Number of rows needs to be > 1.}

\item{method}{Character. Select a unmixing method. Currently, only "NNLS" is implemented. Default is "NNLS".}

\item{iterate}{Integer. Set maximum iteration per pixel. Processing time could increase the more iterations are made possible. Default is 400.}

\item{tolerance}{Numeric. Tolerance limit represneting a nearly zero minimal number. Default is 1e-8.}

\item{...}{further arguments passed to \link[raster]{writeRaster}.}

\item{verbose}{Logical. Prints progress messages during execution.}
}
\value{
RasterBrick. The object will contain one band per endmember, with each value representing the estimated presence propabiliy of the endmember per pixel (0 to 1), and an RMSE band.
}
\description{
\code{mesma} performs a multiple endmember spectral mixture analysis on a multiband raster image. Currently, a sequential coordinate-wise algorithm is used to apply a non negative least square regression (NNLS).
}
\note{
Depending on \code{iterate} and \code{tolerance} settings, the sum of estimated presence propabilites per pixel varies around 1.
}
\examples{

#load packages
library(raster)
library(RStoolbox)

#load an example dataset
data(lsat)

#make up some endmember spectra: water and land
em_names <- c("water", "land")
pts <- data.frame(class=em_names, cell = c(47916,5294))
em <- lsat[pts$cell]
rownames(em) <- em_names

#unmix the image for water and land
props <- mesma(lsat, em, method = "NNLS")

#take a look
raster::hist(props$water)
raster::plot(props$water, col = c("white","blue"))
raster::hist(props$land)
raster::plot(props$land, col = c("white","brown"))

}
\author{
Jakob Schwalb-Willmann
}
