% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normImage.R
\name{normImage}
\alias{normImage}
\title{Center and/or Normalize Raster Images}
\usage{
normImage(img, norm = TRUE, ...)
}
\arguments{
\item{img}{Raster*. Image to transform. Transformation will be performed separately for each layer.}

\item{norm}{Logical. Perform normalization in addition to centering.}

\item{...}{further arguments passed to writeRaster.}
}
\value{
Returns a Raster* with the same number layers as input layers with each layer being centered and optionally normalized.
}
\description{
Normalize / Center an image or image stack.
}
\examples{
data(rlogo)
# Normalization
normImage(rlogo)
# Centering
normImage(rlogo, norm = FALSE)
}

