% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/panSharpen.R
\name{panSharpen}
\alias{panSharpen}
\title{Pan sharpen imagery / Image fusion}
\usage{
panSharpen(img, pan, r, g, b, pc = 1, method = "brovey", norm = TRUE)
}
\arguments{
\item{img}{Raster* object. Coarse resolution imagery}

\item{pan}{RasterLayer. High resolution image, typically panchromatic.}

\item{r}{Character or Integer. Red band in \code{img}. Only relevant if \code{method!='pca'}}

\item{g}{Character or Integer. Green band in \code{img}. Only relevant if \code{method!='pca'}}

\item{b}{Character or Integer. Blue band in \code{img}. Only relevant if \code{method!='pca'}}

\item{pc}{Integer. Which principal component to replace. Usually this should be the first component (default). Only if the first component is dominated by something else than brightness it might be worth a try to use the second component.}

\item{method}{Character. Choose method from c("pca", "ihs", "brovey").}

\item{norm}{Logical. Normalize pan image to 1st PC component. If \code{FALSE} pan will be histogram matched to the 1st PC. Otherwise only min and max are matched.}
}
\description{
provides different methods for pan sharpening a coarse resolution (typically multispectral) image with
a higher reolution panchromatic image. Values of the pan-chromatic and multispectral images must be of the same scale, (e.g. from 0:1, or all DNs from 0:255)
}
\details{
Pan sharpening options:
\itemize{
 \item{\code{method='pca'}: Performs a pca using \link{rasterPCA}. The first component is then swapped for the pan band an the PCA is rotated backwards.}
 \item{\code{method='ihs'}: Performs a color space transform to Intensity-Hue-Saturation space, swaps intensity for the histogram matched pan and does the backwards transformation.}
	\item{\code{method='brovey'}: Performs Brovey reweighting. Pan and img must be at the same value scale (e.g. 0:1, or 0:255) otherwise you'll end up with psychodelic colors.}
}
}

