% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pifMatch.R
\name{pifMatch}
\alias{pifMatch}
\title{PIF-based Image Matching}
\usage{
pifMatch(img, ref, method = "cor", quantile = 0.95)
}
\arguments{
\item{img}{Raster*. Image to be adjusted.}

\item{ref}{Raster*. Reference image.}

\item{method}{Method to calculate pixel similariry}

\item{quantile}{Numeric. Threshold quantile used to identify PIFs}
}
\description{
Automatically estimates pseudo-invariant-features (PIF) by pixel similarity and fits a linear model which is then applied
to the whole scene.
}

