% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readEE.R
\name{readEE}
\alias{readEE}
\title{Tidy import tool for EarthExplorer .csv export files}
\usage{
readEE(x)
}
\arguments{
\item{x}{Character, Character or list. One or more paths to EarthExplorer export files.}
}
\description{
Tidy import tool for EarthExplorer .csv export files
}

