% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageRegistration.R
\name{registerImages}
\alias{registerImages}
\title{Image to Image Registration based on Mutual Information}
\usage{
registerImages(slave, master, shift = 3, shiftInc = 1, n = 500)
}
\arguments{
\item{slave}{Raster* object. Slave image to shift to master.}

\item{master}{Raster* object. Reference image.}

\item{shift}{Numeric or matrix. If numeric, then shift is the maximal absolute radius (in pixels) which \code{slave} is shifted (\code{seq(-shift, shift, by=shiftInc)}).
If shift is a matrix it must have two columns (x shift an y shift), then only these shift values will be tested.}

\item{shiftInc}{Numeric. Shift increment (in pixels, but not restricted to integer). Ignored if \code{shift} is a matrix.}

\item{n}{Integer. Number of samples to calculate mutual information.}
}
\description{
Shifts a slave image to match the reference image (master). Match is based on maximum
mutual information.
}

