% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{data}
\name{rlogo}
\alias{rlogo}
\title{Rlogo as RasterBrick}
\usage{
data(rlogo)
}
\description{
Tiny example of raster data used to run examples.
}
\examples{
data(rlogo)
ggRGB(rlogo,r = 1,g = 2,b = 3)
}
\keyword{datasets}
