% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{data}
\name{rlogo_rs}
\alias{rlogo_rs}
\title{Rlogo as SpatRaster}
\usage{
rlogo_rs
}
\description{
Tiny example of raster data used to run examples.
}
\examples{
ggRGB(rlogo_rs,r = 1,g = 2,b = 3)
}
\keyword{datasets}
