% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rstbxOpts.R
\name{rstbxOpts}
\alias{rstbxOpts}
\title{Set global options for RStoolbox}
\usage{
rstbxOpts(verbose)
}
\arguments{
\item{verbose}{Logical. Make all functions print status messages}
}
\description{
shortcut to options(RStoolbox.*)
}
\examples{
\dontrun{
rstbxOpts(verbose=TRUE)
}
}

