% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{package}
\name{RStoolbox}
\alias{RStoolbox}
\alias{RStoolbox-package}
\title{RStoolbox: A Collection of Remote Sensing Tools}
\description{
RStoolbox: A Collection of Remote Sensing Tools
}

