% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{sam}
\alias{sam}
\title{Spectral Angle Mapper}
\usage{
sam(x, em, angles = FALSE)
}
\arguments{
\item{x}{RasterBrick or RasterStack. Remote sensing imagery (usually hyperspectral)}

\item{em}{Matrix containing endmembers in rows and bands in columns}

\item{angles}{Logical. If \code{TRUE} a RasterBrick containing each one layer per endmember will be returned containing the spectral angles.}
}
\value{
RasterBrick or RasterLayer
If \code{FALSE} a single Layer will be returned in which each pixel is assigned to the closest endmember class.
}
\description{
Spectral Angle Mapper
}

