% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox-package.R
\docType{data}
\name{sen2}
\alias{sen2}
\title{Sentinel 2 MSI L2A Scene}
\usage{
sen2
}
\description{
Contains all 13 bands in already converted spectral reflectances
}
\examples{
ggRGB(sen2, r=4, g=3, b=2, stretch = "lin")
}
\keyword{datasets}
