% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.R
\name{specSimMap}
\alias{specSimMap}
\title{Spectral Angle Mapper}
\usage{
specSimMap(x, em, method = "sam")
}
\arguments{
\item{x}{RasterBrick or RasterStack. Imagery (usually hyperspectral)}

\item{em}{Matrix containing endmembers in rows and bands in columns}
}
\description{
Spectral Angle Mapper
}

