% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectralIndices.R
\name{spectralIndices}
\alias{spectralIndices}
\title{Spectral indices}
\usage{
spectralIndices(inputRaster, blue = NULL, red = NULL, nir = NULL,
  mir = NULL, indices = NULL, L = 0.5, ...)
}
\arguments{
\item{inputRaster}{Raster* object. Typically remote sensing imagery, which is to be classified.}

\item{blue}{Character or integer. Blue band.}

\item{red}{Character or integer. Red band.}

\item{nir}{Character or integer. Near-infrared band.}

\item{mir}{Character or integer. Midwave-infrared band.}

\item{indices}{Character or NULL. One or more spectral indices to calculate (see Details). NULL will calculate all implemented indices given the spectral bands which are provided,}

\item{L}{soil brightness factor for SAVI (0.5 by default)}

\item{...}{further arguments such as filename etc. passed to \link[raster]{writeRaster}}
}
\value{
rasterBrick
}
\description{
Calculate a suite of multispectral indices such as NDVI, SAVI etc. in an efficient way.
}
\examples{
r <- raster(ncol=10,nrow=10)
r[] <- sample(1:155, 100, TRUE)
r <- stack(r, r + 90 + rnorm(100, 10))
names(r) <- c("red", "nir")
SI <- spectralIndices(r, red = 1, nir = 2)
plot(SI)
}
\seealso{
\code{\link[raster]{overlay}}
}

