% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{data}
\name{srtm}
\alias{srtm}
\title{SRTM Digital Elevation Model}
\usage{
data(srtm)
}
\description{
DEM for the Landsat example area taken from SRTM v3 tile: s04_w050_1arc_v3.tif
}
\keyword{datasets}

