% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{data}
\name{srtm_rs}
\alias{srtm_rs}
\title{SRTM Digital Elevation Model}
\usage{
srtm_rs
}
\description{
DEM for the Landsat example area taken from SRTM v3 tile: s04_w050_1arc_v3.tif
}
\examples{
ggR(srtm_rs)
}
\keyword{datasets}
