% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStoolbox-package.R
\docType{data}
\name{srtm_sen2}
\alias{srtm_sen2}
\title{SRTM scene for the sen2 exemplary scene}
\usage{
srtm_sen2
}
\description{
DEM for the Sentinel 2 example area taken from SRTM v4
}
\examples{
ggR(srtm_sen2)
}
\keyword{datasets}
