% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/superClass.R
\name{superClass}
\alias{superClass}
\title{Supervised Classification}
\usage{
superClass(inputRaster, trainingData, classAttributes = NULL,
  nSamples = 100, filename = NULL, maskRaster = NULL, verbose = FALSE,
  predict = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{inputRaster}{Raster* object. Typically remote sensing imagery, which is to be classified.}

\item{trainingData}{SpatialPolygonsDataFrame containing the training data used to train the classifier.}

\item{classAttributes}{character giving the column in \code{trainingData}, which contains the class attribute. Can be omitted, when \code{trainingData} has only one column.}

\item{nSamples}{number of samples per land cover class}

\item{filename}{path to output file (optional). If \code{NULL}, standard raster handling will apply, i.e. storage either in memory or in the raster temp directory.}

\item{maskRaster}{Raster layer containing a binary mask to exclude areas from prediction.}

\item{verbose}{logical. prints progress, statistics and graphics during execution}

\item{predict}{logical. \code{TRUE} (default) will return a classified map, \code{FALSE} will only train the classifier}

\item{...}{further arguments to be passed to randomForest}
}
\value{
A list containing [[1]] the model, [[2]] the predicted raster and [[3]] the class mapping
}
\description{
Supervised Classification
}
\seealso{
\code{\link{randomForest}}
}

