% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superClass.R
\name{validateMap}
\alias{validateMap}
\title{Map accuracy assessment}
\usage{
validateMap(map, valData, responseCol, nSamples = 500,
  mode = "classification", classMapping = NULL)
}
\arguments{
\item{map}{RasterLayer. The classified map.}

\item{valData}{SpatialPolygonsDataFrame or SpatialPointsDataFrame with validation data.}

\item{responseCol}{Character. Column containing the validation data in attribute table of \code{valData}.}

\item{nSamples}{Integer. Number of pixels to sample for validation (only applies to polygons).}

\item{mode}{Character. Either 'classification' or 'regression'.}

\item{classMapping}{optional data.frame with columns \code{'class'} and \code{'classID'} defining the mapping from raster integers to class names.}
}
\description{
validate a map from a classification or regression model. This can be useful to update the accuracy assesment after filtering, e.g. for a minimum mapping unit.
}
\note{
Polygons, which are smaller than the map resolution will only be considered if they overlap with a pixel center coordinate, else they will be ignored
}

