% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifyQA.R
\name{classifyQA}
\alias{classifyQA}
\title{Classify Landsat8 QA Band}
\usage{
classifyQA(img, ...)
}
\arguments{
\item{img}{RasterLayer. Landsat 8 OLI QA band.}

\item{...}{further arguments passed to \link[raster]{writeRaster}}
}
\value{
Returns a RasterLayer with five classes:
\tabular{rr}{
class \tab value \cr
background \tab 1L \cr
cloud  \tab 2L \cr
cirrus \tab 3L \cr
snow   \tab 4L \cr
water  \tab 5L \cr
}
}
\description{
extracts five classes from QA band: background, cloud, cirrus, snow and water.
}
\seealso{
\link{encodeQA} \link{decodeQA}
}

