% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cloudMask.R
\name{cloudShadowMask}
\alias{cloudShadowMask}
\title{Cloud Shadow Detection}
\usage{
cloudShadowMask(img, cm, nc = 5, shiftEstimate = NULL,
  preciseShift = NULL, threshold = 0.1, returnShift = FALSE)
}
\arguments{
\item{img}{Raster* Object containing the scene}

\item{cm}{RasterBrick. Cloud mask (typically the result of \code{cloudMask})}

\item{nc}{Integer. Number of control points. A few points (default) are fine because the final shift is estimated by \link{coregisterImages}.}

\item{shiftEstimate}{NULL or numeric vector of length two (x,y). Estimated displacement of shadows in map units. If \code{NULL}, the user will be asked to select control points interactively.}

\item{preciseShift}{NULL or numeric vector of length two (x,y). Use this if cloud/cloud-shadow displacement is already known, e.g. from a previous run of \code{cloudShadowMask}.}

\item{threshold}{Numeric (between 0 and 1). Threshold used for image co-registration (see details). Doesn't usually require tweaking.}

\item{returnShift}{Logical. Return a numeric vector containing the shift parameters. Usefull if you estimate parameters on a subset of the image.}
}
\description{
Intended for interactive use \code{cloudShadowMask} will ask the user to select a few
corresponding cloud/cloudShadow pixels which will be used to estimate coordinates
for a linear cloudmask shift.
}
\details{
This is a very simplistic approach to cloud shadow masking (simple shift of the cloud mask). It is not image based and accuracy will suffer from clouds at different altitudes. However, just as cloudMask
this is a quick and easy to use tool for Landsat data if you're just working on a few scenes and don't have fMask or CDR data at hand. Although for some test scenes
it does perform surprisingly well.
}
\seealso{
\link{cloudMask}
}

