% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coregisterImages.R
\name{coregisterImages}
\alias{coregisterImages}
\title{Image to Image Co-Registration based on Mutual Information}
\usage{
coregisterImages(slave, master, shift = 3, shiftInc = 1, n = 500,
  doShift = TRUE)
}
\arguments{
\item{slave}{Raster* object. Slave image to shift to master.}

\item{master}{Raster* object. Reference image.}

\item{shift}{Numeric or matrix. If numeric, then shift is the maximal absolute radius (in pixels of \code{master} resolution) which \code{slave} is shifted (\code{seq(-shift, shift, by=shiftInc)}).
If shift is a matrix it must have two columns (x shift an y shift), then only these shift values will be tested.}

\item{shiftInc}{Numeric. Shift increment (in pixels, but not restricted to integer). Ignored if \code{shift} is a matrix.}

\item{n}{Integer. Number of samples to calculate mutual information.}

\item{doShift}{Logical. Perform shift with identifed parameters. If \code{FALSE} it will return the shifts with corresponing mutual information.}
}
\value{
\code{doShift=TRUE} returns a Raster* object (x-y shifted slave image).  \code{doShift=FALSE} returns a data.frame containing the shifts (columns x and y) and the corresponding mutual information (mi).
}
\description{
Shifts a slave image to match the reference image (master). Match is based on maximum
mutual information.
}
\details{
Currently only a simple linear x - y shift is considered and tested. No higher order shifts (e.g. rotation, non-linear transformation) are performed. This means that your imagery
should already be properly geometrically corrected.
}

