% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodeQA.R
\name{decodeQA}
\alias{decodeQA}
\title{Decode QA flags to bit-words}
\usage{
decodeQA(x)
}
\arguments{
\item{x}{Integer (16bit)}
}
\description{
Intended for use with the Landsat 8 OLI QA band. Decodes pixel quality flags from integer to bit-words.
}
\examples{
decodeQA(53248)
}
\seealso{
\link{encodeQA}
}

