% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeQA.R
\name{decodeQA}
\alias{decodeQA}
\title{Decode QA flags to bit-words}
\usage{
decodeQA(x)
}
\arguments{
\item{x}{Integer (16bit)}
}
\description{
Intended for use with the Landsat 8 OLI QA band. Decodes pixel quality flags from integer to bit-words.
NOTE: THE NEW USGS LANDSAT COLLECTION 1 DATA HAS CHANGED THE QA BITWORD ASSIGNMENTS. THIS IS NOT YET IMPLEMENTED HERE.
DO NOT USE encodeQA, decodeQA and classifyQA FOR LANDSAT COLLECTION 1 DATA (yet)!
}
\examples{
decodeQA(53248)
}
\seealso{
\link{encodeQA}
}
