% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodeQA.R
\name{decodeQA}
\alias{decodeQA}
\title{Decode QA flags to bit-words}
\usage{
decodeQA(x)
}
\arguments{
\item{x}{Integer (16bit)}
}
\description{
Decode QA flags to bit-words
}
\examples{
decodeQA(53248)
}

