% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimateSHV}
\alias{estimateSHV}
\title{Estimate image haze for dark object subtraction procedures}
\usage{
estimateSHV(x, band, darkProp = 0.001, plot = FALSE, returnTable = FALSE)
}
\arguments{
\item{x}{raster object from which to estimate haze}

\item{band}{character. Band or bandname from which to estimate SHV (optinal if x contains only one layer)}

\item{darkProp}{proportion of pixels estimated to be dark}

\item{plot}{display histograms and haze values}

\item{returnTable}{return the table per layer}
}
\description{
Estimate image haze for dark object subtraction procedures
}

