% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggR.R
\name{ggR}
\alias{ggR}
\title{Plot rasters in ggplot with greyscale}
\usage{
ggR(x, layer = 1, maxpixels = 5e+06, lowColor = "white",
  highColor = "black", legendName = "Legend", ggObj = TRUE)
}
\arguments{
\item{x}{raster}

\item{layer}{layername}

\item{maxpixels}{Integer. Maximal number of pixels to sample}

\item{lowColor}{Character. Color for lowest value}

\item{highColor}{Character. Color for highest value}

\item{legendName}{Character. Layer name}

\item{ggObj}{Logical. Return a ggplot2 object (TRUE) or just the data.frame}
}
\description{
Plot rasters in ggplot with greyscale
}

