% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggRGB}
\alias{ggRGB}
\title{Create ggplot raster with RGB from 3 rasterLayers}
\usage{
ggRGB(x, r = 3, g = 2, b = 1, scale, maxpixels = 5e+05,
  stretch = NULL, ext = NULL, minMax = NULL, clipToMinMax = FALSE,
  ggObj = TRUE, coordEqual = TRUE, interpolate = TRUE,
  annotation = TRUE)
}
\arguments{
\item{x}{RasterStack or RasterBrick}

\item{r}{Integer or character. Red layer in x}

\item{g}{Integer or character. Green layer in x}

\item{b}{Integer or character. Blue layer in x}

\item{scale}{Numeric. Maximum possible pixel value (optional). Defaults to 255 or to the maximum value of x if that is larger than 255}

\item{maxpixels}{Integer. Maximal number of pixels used for plotting.}

\item{stretch}{Character. Either 'lin' or 'hist' for linear or histogram stretch of the data}

\item{ext}{extent object tp crop the image}

\item{minMax}{Vector or matrix. Can be used to reduce the range of values. Either a vector of two values for all bands (c(min, max)) or
a 3x2 matrix with separate min and max values (columns) for each layer (rows).}

\item{clipToMinMax}{Logical. If \code{TRUE}, values > scale will be set to NA. if \code{FALSE} they will be set to scale. Defaults to \code{FALSE}.}

\item{ggObj}{Logical. If \code{TRUE} a ggplot2 object is returned. If \code{FALSE} a data.frame with coordinates and color will be returned.}

\item{coordEqual}{Logical. Uses coord_equal, i.e. aspect ratio of 1:1.}

\item{interpolate}{Logical. Interpolate the raster during plotting.}

\item{annotation}{Logical. If \code{TRUE} annotation_raster is used, otherwise geom_raster()+scale_fill_identity is used. Note that you can't use scale_fill* in addition to the latter, because it alread requires scale_fill_identity().}
}
\value{
A ggplot2 object, or a three column data frame with coordinates and fill colour.
}
\description{
Calculates RGB color composite raster for plotting with ggplot2. Optional value clipping and and stretching can be used to enhance the imagery.
}
\details{
Functionality is based on \code{\link[raster]{plotRGB}} from the raster package.
}
\examples{
br <- brick(system.file("external/rlogo.grd", package="raster"))
ggRGB(br, 1, 2, 3)
ggRGB(br, r=1,g=2,b=3, minMax = matrix(c(100,150,10,200,50,255),  ncol = 2, by = TRUE))
}

