% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/histMatch.R
\name{histMatch}
\alias{histMatch}
\title{Image to Image Contrast Matching}
\usage{
histMatch(x, ref, nsamp = 1e+05, intersectOnly = TRUE, precise = TRUE,
  ...)
}
\arguments{
\item{x}{RasterLayer. Source raster which is to be modified.}

\item{ref}{RasterLayer. Reference raster, to which x will be matched.}

\item{nsamp}{integer. Number of random samples to build the histograms.}

\item{intersectOnly}{logical. If \code{TRUE} sampling will only take place in the overlap extent of the two rasters. Otherwise the full rasters will be used for sampling.}

\item{precise}{logical. If \code{TRUE} the exact same pixels will be used in the overlap.}

\item{...}{Further arguments to be passed to \link[raster]{writeRaster}.}
}
\value{
A RasterLayer of x adjusted to the histogram of ref.
}
\description{
Performs image to image contrast adjustments based on histogram matching using empirical cumulative distribution functions from both images.
}
\examples{
par(mfrow=c(1,3))
r0  <- raster(system.file("external/test.grd", package="raster"))
plot(r0, main = "original meuse dataset")

r1 <- crop(r0, extent(c(180000,182000,330000,334000))) * 10
r.merge <- merge(r1, r0)
plot(r.merge, main ="merge r0+r1 without histMatch")

r.match <- histMatch(r0, r1)
rcmatch <- merge(r1, r.match)
plot(rcmatch, main = "merge r0+r1 with histMatch")
}

