% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/radCor.R
\docType{data}
\name{LANDSAT.db}
\alias{LANDSAT.db}
\title{Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php}
\format{\preformatted{List of 3
 $ LANDSAT5:List of 1
  ..$ TM:'data.frame':	7 obs. of  26 variables:
  .. ..$ band           : chr [1:7] "B1_DN" "B2_DN" "B3_DN" "B4_DN" ...
  .. ..$ bandtype       : chr [1:7] "REF" "REF" "REF" "REF" ...
  .. ..$ centerWavl     : num [1:7] 0.485 0.569 0.66 0.84 1.676 ...
  .. ..$ spatRes1       : num [1:7] 30 30 30 30 30 30 30
  .. ..$ spatRes2       : num [1:7] 30 30 30 30 30 30 60
  .. ..$ esun           : num [1:7] 1983 1796 1536 1031 220 ...
  .. ..$ B1_DN_veryClear: num [1:7] 1 0.52786 0.2916 0.11113 0.00701 ...
  .. ..$ B1_DN_clear    : num [1:7] 1 0.7265 0.54 0.3334 0.0837 ...
  .. ..$ B1_DN_moderate : num [1:7] 1 0.852 0.735 0.577 0.289 ...
  .. ..$ B1_DN_hazy     : num [1:7] 1 0.894 0.806 0.681 0.42 ...
  .. ..$ B1_DN_veryHazy : num [1:7] 1 0.923 0.857 0.76 0.538 ...
  .. ..$ B2_DN_veryClear: num [1:7] 1.8944 1 0.5524 0.2105 0.0133 ...
  .. ..$ B2_DN_clear    : num [1:7] 1.376 1 0.743 0.459 0.115 ...
  .. ..$ B2_DN_moderate : num [1:7] 1.173 1 0.862 0.677 0.339 ...
  .. ..$ B2_DN_hazy     : num [1:7] 1.118 1 0.901 0.761 0.469 ...
  .. ..$ B2_DN_veryHazy : num [1:7] 1.083 1 0.929 0.823 0.583 ...
  .. ..$ B3_DN_veryClear: num [1:7] 3.429 1.81 1 0.381 0.024 ...
  .. ..$ B3_DN_clear    : num [1:7] 1.852 1.345 1 0.617 0.155 ...
  .. ..$ B3_DN_moderate : num [1:7] 1.361 1.16 1 0.786 0.394 ...
  .. ..$ B3_DN_hazy     : num [1:7] 1.241 1.109 1 0.845 0.521 ...
  .. ..$ B3_DN_veryHazy : num [1:7] 1.167 1.077 1 0.886 0.628 ...
  .. ..$ B4_DN_veryClear: num [1:7] 8.9981 4.7497 2.6239 1 0.0631 ...
  .. ..$ B4_DN_clear    : num [1:7] 3 2.179 1.62 1 0.251 ...
  .. ..$ B4_DN_moderate : num [1:7] 1.732 1.476 1.273 1 0.501 ...
  .. ..$ B4_DN_hazy     : num [1:7] 1.469 1.313 1.184 1 0.617 ...
  .. ..$ B4_DN_veryHazy : num [1:7] 1.316 1.215 1.128 1 0.708 ...
 $ LANDSAT7:List of 1
  ..$ ETM:'data.frame':	8 obs. of  26 variables:
  .. ..$ band           : chr [1:8] "B1_DN" "B2_DN" "B3_DN" "B4_DN" ...
  .. ..$ bandtype       : chr [1:8] "REF" "REF" "REF" "REF" ...
  .. ..$ spatRes1       : num [1:8] 30 30 30 30 30 30 30 15
  .. ..$ spatRes2       : num [1:8] 30 30 30 30 30 30 60 15
  .. ..$ centerWavl     : num [1:8] 0.485 0.56 0.66 0.835 1.65 ...
  .. ..$ esun           : num [1:8] 1997 1812 1533 1039 231 ...
  .. ..$ B1_DN_veryClear: num [1:8] 1 0.56262 0.2916 0.11382 0.00747 ...
  .. ..$ B1_DN_clear    : num [1:8] 1 0.7501 0.54 0.3374 0.0864 ...
  .. ..$ B1_DN_moderate : num [1:8] 1 0.866 0.735 0.581 0.294 ...
  .. ..$ B1_DN_hazy     : num [1:8] 1 0.904 0.806 0.684 0.424 ...
  .. ..$ B1_DN_veryHazy : num [1:8] 1 0.931 0.857 0.762 0.542 ...
  .. ..$ B2_DN_veryClear: num [1:8] 1.7774 1 0.5183 0.2023 0.0133 ...
  .. ..$ B2_DN_clear    : num [1:8] 1.333 1 0.72 0.45 0.115 ...
  .. ..$ B2_DN_moderate : num [1:8] 1.155 1 0.848 0.671 0.339 ...
  .. ..$ B2_DN_hazy     : num [1:8] 1.106 1 0.891 0.756 0.469 ...
  .. ..$ B2_DN_veryHazy : num [1:8] 1.075 1 0.921 0.819 0.583 ...
  .. ..$ B3_DN_veryClear: num [1:8] 3.4293 1.9294 1 0.3903 0.0256 ...
  .. ..$ B3_DN_clear    : num [1:8] 1.852 1.389 1 0.625 0.16 ...
  .. ..$ B3_DN_moderate : num [1:8] 1.36 1.18 1 0.79 0.4 ...
  .. ..$ B3_DN_hazy     : num [1:8] 1.241 1.122 1 0.848 0.527 ...
  .. ..$ B3_DN_veryHazy : num [1:8] 1.167 1.086 1 0.889 0.632 ...
  .. ..$ B4_DN_veryClear: num [1:8] 8.7858 4.943 2.5619 1 0.0656 ...
  .. ..$ B4_DN_clear    : num [1:8] 2.964 2.223 1.601 1 0.256 ...
  .. ..$ B4_DN_moderate : num [1:8] 1.722 1.491 1.265 1 0.506 ...
  .. ..$ B4_DN_hazy     : num [1:8] 1.463 1.323 1.179 1 0.621 ...
  .. ..$ B4_DN_veryHazy : num [1:8] 1.312 1.221 1.125 1 0.711 ...
 $ LANDSAT8:List of 1
  ..$ OLI_TIRS:'data.frame':	12 obs. of  26 variables:
  .. ..$ band           : chr [1:12] "B1_DN" "B2_DN" "B3_DN" "B4_DN" ...
  .. ..$ bandtype       : chr [1:12] "REF" "REF" "REF" "REF" ...
  .. ..$ spatRes1       : num [1:12] 30 30 30 30 30 30 30 30 30 30 ...
  .. ..$ spatRes2       : num [1:12] 30 30 30 30 30 30 30 30 30 30 ...
  .. ..$ centerWavl     : num [1:12] 0.44 0.48 0.56 0.655 0.865 1.61 2.2 1.37 10.6 11.5 ...
  .. ..$ esun           : num [1:12] NA 2067 1893 1603 973 ...
  .. ..$ B1_DN_veryClear: num [1:12] 1 0.7061 0.3811 0.2036 0.0669 ...
  .. ..$ B1_DN_clear    : num [1:12] 1 0.84 0.617 0.451 0.259 ...
  .. ..$ B1_DN_moderate : num [1:12] 1 0.917 0.786 0.672 0.509 ...
  .. ..$ B1_DN_hazy     : num [1:12] 1 0.941 0.845 0.757 0.623 ...
  .. ..$ B1_DN_veryHazy : num [1:12] 1 0.957 0.886 0.82 0.713 ...
  .. ..$ B2_DN_veryClear: num [1:12] 1.4163 1 0.5398 0.2884 0.0948 ...
  .. ..$ B2_DN_clear    : num [1:12] 1.19 1 0.735 0.537 0.308 ...
  .. ..$ B2_DN_moderate : num [1:12] 1.091 1 0.857 0.733 0.555 ...
  .. ..$ B2_DN_hazy     : num [1:12] 1.063 1 0.898 0.804 0.662 ...
  .. ..$ B2_DN_veryHazy : num [1:12] 1.044 1 0.926 0.856 0.745 ...
  .. ..$ B3_DN_veryClear: num [1:12] 2.624 1.853 1 0.534 0.176 ...
  .. ..$ B3_DN_clear    : num [1:12] 1.62 1.361 1 0.731 0.419 ...
  .. ..$ B3_DN_moderate : num [1:12] 1.273 1.167 1 0.855 0.647 ...
  .. ..$ B3_DN_hazy     : num [1:12] 1.184 1.114 1 0.896 0.738 ...
  .. ..$ B3_DN_veryHazy : num [1:12] 1.128 1.08 1 0.925 0.805 ...
  .. ..$ B4_DN_veryClear: num [1:12] 4.911 3.467 1.872 1 0.329 ...
  .. ..$ B4_DN_clear    : num [1:12] 2.216 1.862 1.368 1 0.573 ...
  .. ..$ B4_DN_moderate : num [1:12] 1.489 1.365 1.17 1 0.757 ...
  .. ..$ B4_DN_hazy     : num [1:12] 1.321 1.243 1.116 1 0.823 ...
  .. ..$ B4_DN_veryHazy : num [1:12] 1.22 1.17 1.08 1 0.87 ...
}}
\usage{
LANDSAT.db
}
\description{
Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php
}
\keyword{internal}

