% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{LANDSAT.db}
\alias{LANDSAT.db}
\title{Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php}
\format{\preformatted{List of 3
 $ LANDSAT5:List of 1
  ..$ TM:'data.frame':	7 obs. of  6 variables:
  .. ..$ band      : Factor w/ 7 levels "B1","B2","B3",..: 1 2 3 4 5 6 7
  .. ..$ bandtype  : Factor w/ 2 levels "REF","TIR": 1 1 1 1 1 2 1
  .. ..$ centerWavl: num [1:7] 0.485 0.569 0.66 0.84 1.676 ...
  .. ..$ spatRes1  : num [1:7] 30 30 30 30 30 30 30
  .. ..$ spatRes2  : num [1:7] 30 30 30 30 30 60 30
  .. ..$ esun      : num [1:7] 1983 1796 1536 1031 220 ...
 $ LANDSAT7:List of 1
  ..$ ETM:'data.frame':	8 obs. of  6 variables:
  .. ..$ band      : Factor w/ 8 levels "B1","B2","B3",..: 1 2 3 4 5 6 7 8
  .. ..$ bandtype  : Factor w/ 3 levels "PAN","REF","TIR": 2 2 2 2 2 3 2 1
  .. ..$ spatRes1  : num [1:8] 30 30 30 30 30 30 30 15
  .. ..$ spatRes2  : num [1:8] 30 30 30 30 30 60 30 15
  .. ..$ centerWavl: num [1:8] 0.483 0.56 0.662 0.825 1.648 ...
  .. ..$ esun      : num [1:8] 1997 1812 1533 1039 231 ...
 $ LANDSAT8:List of 1
  ..$ OLI_TIRS:'data.frame':	12 obs. of  6 variables:
  .. ..$ band      : Factor w/ 12 levels "B1","B10","B11",..: 1 4 5 6 7 8 9 10 11 2 ...
  .. ..$ bandtype  : Factor w/ 4 levels "PAN","QA","REF",..: 3 3 3 3 3 3 3 1 3 4 ...
  .. ..$ spatRes1  : num [1:12] 30 30 30 30 30 30 30 15 30 30 ...
  .. ..$ spatRes2  : num [1:12] 30 30 30 30 30 30 30 15 30 30 ...
  .. ..$ centerWavl: num [1:12] 0.44 0.48 0.56 0.655 0.865 1.61 2.2 0.59 1.37 10.6 ...
  .. ..$ esun      : logi [1:12] NA NA NA NA NA NA ...
}}
\usage{
LANDSAT.db
}
\description{
Landsat auxilliary data. Taken from Chander et al 2009
spatRes resampling: http://landsat.usgs.gov/band_designations_landsat_satellites.php
}
\keyword{datasets}

