% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normImage.R
\name{normImage}
\alias{normImage}
\title{Linear image normalization}
\usage{
normImage(x, y, xmin, xmax, ymin, ymax, forceMinMax = FALSE)
}
\arguments{
\item{x}{Raster* object. Image to normalise.}

\item{y}{Raster* object. Reference image.}

\item{xmin}{Numeric. Min value of x.}

\item{xmax}{Numeric. Max value of x.}

\item{ymin}{Numeric. Min value of y.}

\item{ymax}{Numeric. Max value of y.}

\item{forceMinMax}{Logical. Forces update of min and max data slots in x or y.}
}
\description{
performs linear shifts of value ranges either to match another image
or to any other min and max value.
}
\seealso{
\link{histMatch}
}

