% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/panSharpen.R
\name{panSharpen}
\alias{panSharpen}
\title{Pan sharpen imagery / Image fusion}
\usage{
panSharpen(img, pan, r, g, b, method = "pca", norm = TRUE)
}
\arguments{
\item{img}{Raster* object. Coarse resolution imagery}

\item{pan}{RasterLayer. High resolution image, typically panchromatic.}

\item{r}{Character or Integer. Red band in \code{img}. Only relevant if \code{method='pca'}}

\item{g}{Character or Integer. Green band in \code{img}. Only relevant if \code{method='pca'}}

\item{b}{Character or Integer. Blue band in \code{img}. Only relevant if \code{method='pca'}}

\item{method}{Character. Choose method from c("pca", "ihs").}

\item{norm}{Logical. Normalize pan image to 1st PC component. If \code{FALSE} pan will be histogram matched to the 1st PC. Otherwise only min and max are matched.}
}
\description{
provides different methods for pan sharpening a coarse resolution (typically multispectral) image with
a higher reolution panchromatic image
}
\details{
Pan sharpening options:
\itemize{
 \item{\code{method='pca'}: Performs a pca using \link{rasterPCA}. The first component is then swapped for the pan band an the PCA is rotated backwards.}
 \item{\code{method='ihs'}: Performs a color space transform to Intensity-Hue-Saturation space, swaps intensity for the histogram matched pan and does the backwards transformation.}
}
}

