% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unsuperClass.R
\name{predict.kmeans}
\alias{predict.kmeans}
\title{Predict method for kmeans objects}
\usage{
\method{predict}{kmeans}(object, newdata, ...)
}
\arguments{
\item{object}{\link[stats]{kmeans} object}

\item{newdata}{matrix}

\item{...}{further arguments. None implemented.}
}
\description{
Prediction for kmeans models based on minimum distance to cluster centers
}

