% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radCor.R
\name{radCor}
\alias{radCor}
\title{Radiometric calibration and correction}
\usage{
radCor(img, metaData, radianceOnly = FALSE, method = "APREF",
  bandSet = "full", SHV, hazeBand, atHaze, darkProp, verbose = FALSE)
}
\arguments{
\item{img}{raster object}

\item{metaData}{object of class ImgMetaData or a path to the meta data (MTL) file.}

\item{radianceOnly}{Logical. If \code{TRUE} only radiance will be calculated. Otherwise reflectance and brightness temperature are calculated.}

\item{method}{Radiometric correction method to be used. There are currently four methods available (see Details):
"APREF", "SDOS", "DOS", "COSTZ".}

\item{bandSet}{numeric or character. original Landsat band numbers or names in the form of ("B1", "B2" etc). If set to 'full' all bands in the solar region will be processed.}

\item{SHV}{starting haze value, can be estimated using estimateSHV(). if not provided and method is "DOS" or "COSTZ" SHV will be estimated in an automated fashion. Not needed for apparent reflectance.}

\item{hazeBand}{band from which SHV was estimated.}

\item{atHaze}{character. Scene haze characteristics. Will be estimated if not expicilty provided. Must be one of \code{"veryClear", "clear", "moderate", "hazy"} or \code{"veryHazy"}.}

\item{verbose}{Logical. Print status information.}
}
\description{
Implements several different methods for absolute radiometric correction of Landsat data.
You can either specify a metadata file, or supply all neccesary values manually. With proper parametrization APREF and SDOS should work for other sensors as well.
}
\details{
Conversion to top of atmosphere radiance (Watts/(m2 * srad * \eqn{\mu}m))
Conversion to at-satellite brightness temperature (K)
Conversion to top of atmosphere reflectance (unitless) corrected for the sun angle
Estimation of at-surface spectral reflectance (unitless)
}
\note{
This was originally a fork of randcorr in the landsat package. It may be slower, however it works on Raster* objects and hence is memory-safe.
}
\references{
S. Goslee (2011): Analyzing Remote Sensing Data in R: The landsat Package. Journal of Statistical Software 43(4).
}

