% Generated by roxygen2 (4.0.1): do not edit by hand
\name{radCor}
\alias{radCor}
\title{Radiometric calibration and correction}
\usage{
radCor(x, metaData, reflectance = TRUE, satellite, bandSet = "full", gain,
  offset, G_rescale, B_rescale, sunElev, satZenith = 0, edist, esun, date,
  SHV, hazeBand, atHaze, method = "APREF")
}
\arguments{
\item{x}{raster object}

\item{metaData}{either the result of \code{readMeta} or a path to the meta data (MCL) file.}

\item{reflectance}{logical. If \code{TRUE} output will be reflectance, if \code{FALSE} it will be radiance}

\item{bandSet}{numeric or character. original Landsat band numbers or names in the form of ("B1", "B2" etc). If set to 'full' all bands in the solar region will be processed.}

\item{gain}{Band-specific sensor gain. Require either gain and offset or Grescale and Brescale to convert DN to radiance.}

\item{offset}{Band-specific sensor offset. Require either gain and offset or Grescale and Brescale to convert DN to radiance.}

\item{Grescale}{Band-specific sensor Grescale (gain). Require either gain and offset or Grescale and Brescale to convert DN to radiance.}

\item{Brescale}{Band-specific sensor Brescale (bias). Require either gain and offset or Grescale and Brescale to convert DN to radiance.}

\item{sunElev}{Sun elevation in degrees}

\item{satZenith}{Satellite sensor zenith angle (0 for Landsat)}

\item{edist}{Earth-Sun distance in AU.}

\item{esun}{Mean exo-atmospheric solar irradiance, as given by Chandler et al. 2009 or others.}

\item{Lhaze}{Haze value, such as SHV from DOS() function. Not needed for apparent reflectance.}

\item{Radiometric}{correction method to be used. There are currently four methods available:
"APREF", "DOS" (Chavez 1989), "COSTZ" (Chavez 1996).}
}
\description{
Implements several different methods for absolute radiometric correction of Landsat data.
You can either specify a metadata file, or supply all neccesary values manually. With proper parametrization APREF and SDOS should work for other sensors as well.
}
\note{
This is a fork of randcorr in the landsat package. It may be slower, however it works on Raster* objects and hence is memory-safe.
}
\references{
S. Goslee (2011): Analyzing Remote Sensing Data in R: The landsat Package. Journal of Statistical Software 43(4).
}
\seealso{
\link[landsat]{radiocorr}
}

