% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radCor.R
\name{radCor}
\alias{radCor}
\title{Radiometric calibration and correction}
\usage{
radCor(img, metaData, radiance = FALSE, method = "APREF",
  bandSet = "full", SHV, hazeBand, atHaze, darkProp = 0.02, verbose)
}
\arguments{
\item{img}{raster object}

\item{metaData}{object of class ImageMetaData or a path to the meta data (MTL) file.}

\item{radiance}{Logical. If \code{TRUE} only radiance will be calculated. Otherwise reflectance and brightness temperature are calculated.}

\item{method}{Radiometric correction method to be used. There are currently four methods available (see Details):
"APREF", "SDOS", "DOS", "COSTZ".}

\item{bandSet}{numeric or character. original Landsat band numbers or names in the form of ("B1", "B2" etc). If set to 'full' all bands in the solar region will be processed.}

\item{SHV}{starting haze value, can be estimated using estimateSHV(). if not provided and method is "DOS" or "COSTZ" SHV will be estimated in an automated fashion. Not needed for apparent reflectance.}

\item{hazeBand}{band from which SHV was estimated.}

\item{atHaze}{character. Scene haze characteristics. Will be estimated if not expicilty provided. Must be one of \code{"veryClear", "clear", "moderate", "hazy"} or \code{"veryHazy"}.}

\item{darkProp}{numeric. Estimated proportion of dark pixels in the scene. Used only for automatic guessing of SHV.}

\item{verbose}{Logical. Print status information.}
}
\description{
Implements several different methods for absolute radiometric correction of Landsat data.
You can either specify a metadata file, or supply all neccesary values manually. With proper parametrization APREF and SDOS should work for other sensors as well.
}
\details{
Conversion to top of atmosphere radiance (\eqn{W/(m^2 * srad * \mu m)})
Conversion to at-satellite brightness temperature (K)
Conversion to top of atmosphere reflectance (unitless) corrected for the sun angle
Estimation of at-surface spectral reflectance (unitless)

\describe{
\item{APREF}{Apparent reflectance}
\item{DOS}{Dark object subtratction following Chavez (1989)}
\item{COSTZ}{Dark object subtraction following Chaves(1996)}
\item{SDOS}{Simple dark object subtraction. Classical DOS, Lhaze must be estimated for each band separately.}
}

The implemented sun-earth distances neglect the earth's eccentricity. Instead it uses a 100 year daily average (1979-2070).


Atmospheric haze decay model according to Chavez (1989)
\describe{
\item{veryClear}{\eqn{\lambda^{-4.0}}}
\item{clear}{\eqn{\lambda^{-2.0}}}
\item{moderate}{\eqn{\lambda^{-1.0}}}
\item{hazy}{\eqn{\lambda^{-0.7}}}
\item{veryHazy}{\eqn{\lambda^{-0.5}}}
}
}
\note{
This was originally a fork of randcorr in the landsat package. It may be slower, however it works on Raster* objects and hence is memory-safe.
}
\references{
S. Goslee (2011): Analyzing Remote Sensing Data in R: The landsat Package. Journal of Statistical Software 43(4).
}

