% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/changeVectors.R
\name{rasterCVA}
\alias{rasterCVA}
\title{Change Vector Analysis}
\usage{
rasterCVA(x, y, ...)
}
\arguments{
\item{x}{RasterBrick or RasterStack with two layers.}

\item{y}{RasterBrick or RasterStack with two layers.}

\item{...}{further arguments passed to writeRaster}
}
\description{
Calculateds angle and magnitude of change vectors.
Dimensionality is limited to two bands.
}

