% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readMeta}
\alias{readMeta}
\title{Read landsat MTL metadata files}
\usage{
readMeta(file, unifiedMetadata = TRUE, unifiedOnly = FALSE)
}
\arguments{
\item{file}{path to Landsat MTL file (...MTL.txt)}

\item{unifiedMetadata}{Logical. If \code{TRUE} some relevant etadata of Landsat 5:8 are homogenized into a standard format and appended to the original metadata.}

\item{unifiedOnly}{Logical. Return only unified (pretty) metadata.}
}
\value{
Returns a list containing the Metadata of the MTL or XML file, structured by the original grouping.
}
\description{
Besides reading metadata, readMeta deals with legacy versions of Landsat metadata files and where possible adds missing information (radiometric gain and offset, earth-sun distance).
}

