% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readMeta.R
\name{readMeta}
\alias{readMeta}
\title{Read landsat MTL metadata files}
\usage{
readMeta(file, raw = FALSE)
}
\arguments{
\item{file}{path to Landsat MTL file (...MTL.txt)}

\item{raw}{Logical. If \code{TRUE} the full raw metadate will be returned as a list. All important metadata are homogenized into a standard format (ImageMetaData) and some information is added.}
}
\value{
Object of class ImageMetaData
}
\description{
Besides reading metadata, readMeta deals with legacy versions of Landsat metadata files and where possible adds missing information (radiometric gain and offset, earth-sun distance).
}

