% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readSLI}
\alias{readSLI}
\title{Read ENVI spectral libraries}
\usage{
readSLI(path)
}
\arguments{
\item{path}{path to spectral library file with ending .sli.}
}
\description{
ENVI spectral libraries consist of a binary data file (.sli) and a corresponding header (.hdr, or .sli.hdr) file. Spectra are read into a data.frame with the columns:
Wavelength, Spectrum 1, Spectrum 2 ...
}
\examples{
## Create fake spectra
data <- data.frame(wavelength=350:2500, spectrumA=cumsum(abs(rnorm(2151))), spectrumB=cumsum(abs(rnorm(2151))))
pathToFile <- str_c(tempdir(),"/specLib.sli")

## Write to binary spectral library
writeSLI(x = data, path = pathToFile)

## Read from binary spectral library
dataRe <- readSLI(path = pathToFile)

## Check whether they are the same
all.equal(data, dataRe)
## oops? nope, only colnames mismatch. numbers are fine.
cat("colnames:\\noriginal: ",colnames(data),"\\nre-read: ",colnames(dataRe))
}
\references{
\href{http://geol.hu/data/online_help/ENVI_Header_Format.html}{ENVI header format}
}
\seealso{
\code{\link{writeSLI}}
}

