% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rsOpts.R
\name{rsOpts}
\alias{rsOpts}
\title{Set global options for RStoolbox}
\usage{
rsOpts(verbose)
}
\arguments{
\item{verbose}{Logical. Make all functions print status messages}
}
\description{
shortcut to options(RStoolbox.*)
}
\examples{
\dontrun{
rsOpts(verbose=TRUE)
}
}

