% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsOpts.R
\name{rsOpts}
\alias{rsOpts}
\title{Set global options for RStoolbox}
\usage{
rsOpts(verbose)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE} many functions will print status messages about the current processing step. By default verbose mode is disabled.}
}
\description{
shortcut to options(RStoolbox.*)
}
\examples{
# rsOpts(verbose=TRUE)
}

