% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RStoolbox.R
\docType{package}
\name{RStoolbox}
\alias{RStoolbox}
\alias{RStoolbox-package}
\title{RStoolbox: A Collection of Remote Sensing Tools}
\description{
The RStoolbox package provides a set of functions which simplify performing standard remote sensing tasks in R.
}
\section{Data import}{


\itemize{
 \item{readMeta}{import Landsat metadata from MTL or XML files}
 \item{stackMeta}{load Landsat bands based on metadata}
}
}

\section{Data Pre-Processing}{


\itemize{
 \item{radCor}{radiometric conversions and corrections. Primarily, yet not exclusively, intended for Landsat data processing. DN to radiance to reflectance conversion as well as DOS approaches}
 \item{cloudMask}{mask clouds in Landsat or other imagery which comes with a thermal band}
 \item{histMatch}{matches the histograms of two scenes}
}
}

\section{Data Analysis}{


\itemize{
\item{spectralIndices}:{ calculate a set of predefined multispectral indices like NDVI}
\item{unsuperClass}{unsupervised classification}
\item{superClass}{supervised classification}
\item{fCover}{fractional cover of coarse resolution imagery based on high resolution classificaton}
}
}

\section{Data Display}{


\itemize{
\item{ggRGB}{efficient plotting of remote sensing imagery in RGB with ggplot2}
}
}

