% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{spectralIndices}
\alias{spectralIndices}
\title{Spectral indices}
\usage{
spectralIndices(inputRaster, indices = "NDVI", sensor, bands,
  maskRaster = NULL, verbose = FALSE, ...)
}
\arguments{
\item{inputRaster}{Raster* object. Typically remote sensing imagery, which is to be classified.}

\item{indices}{character. one or more spectral indices}

\item{sensor}{if a sensor is specified \code{bands} is populated automatically. Specifying a sensor requires the layernames in inputRaster to match the official band designations formatted as "B1", "B2" etc.}

\item{bands}{list of band designations. See notes for details}

\item{maskRaster}{Raster layer containing a binary mask to exclude areas from prediction.}

\item{verbose}{logical. prints progress, statistics and graphics during execution}

\item{...}{further arguments such as filename etc. passed to \link[raster]{writeRaster}}
}
\value{
rasterBrick or rasterStack
}
\description{
Spectral indices
}
\examples{
r <- raster(ncol=10,nrow=10)
r[] <- sample(1:155, 100, TRUE)
r <- stack(r, r + 90 + rnorm(100, 10))
names(r) <- c("red", "nir")
SI <- spectralIndices(r, indices = c("SR", "NDVI"), bands = list(NIR = "nir", RED = "red"))
plot(SI)
}
\seealso{
\code{\link[raster]{overlay}}
}

