% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/superClass.R
\name{superClass}
\alias{superClass}
\title{Supervised Classification}
\usage{
superClass(inputRaster, trainingData, responseCol = NULL, nSamples = 100,
  model = "rf", tuneLength = 3, filename = NULL, verbose = FALSE,
  predict = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{inputRaster}{Raster* object. Typically remote sensing imagery, which is to be classified.}

\item{trainingData}{SpatialPolygonsDataFrame containing the training data used to train the classifier.}

\item{responseCol}{character giving the column in \code{trainingData}, which contains the response variable. Can be omitted, when \code{trainingData} has only one column.}

\item{nSamples}{number of samples per land cover class}

\item{model}{Character. Which model to use. See \link[caret]{train} for options. Defaults to randomForest ('rf')}

\item{tuneLength}{Integer. Number of levels for each tuning parameters that should be generated by \link[caret]{train}.}

\item{filename}{path to output file (optional). If \code{NULL}, standard raster handling will apply, i.e. storage either in memory or in the raster temp directory.}

\item{verbose}{logical. prints progress, statistics and graphics during execution}

\item{predict}{logical. \code{TRUE} (default) will return a classified map, \code{FALSE} will only train the classifier}

\item{...}{further arguments to be passed to caret}
}
\value{
A list containing [[1]] the model, [[2]] the predicted raster and [[3]] the class mapping
}
\description{
Supervised classification both for classification and regression mode based on vector training data (points or polygons).
}
\seealso{
\code{\link[caret]{train}}
}

