% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topCor.R
\name{topCor}
\alias{topCor}
\title{Topographic Illumination Correction}
\usage{
topCor(img, dem, metaData, method = "cos", stratImg, stratMeth = "noStrat",
  stratN = 5, cosi)
}
\arguments{
\item{img}{Raster*. Imagery to correct}

\item{dem}{Raster*. Either a digital elevation model as a RasterLayer.}

\item{metaData}{Character, ImageMetaData or List. Either a path to a meta-data file, ImageMetaData object (see \link{readMeta}) or a numeric vector containing sun azimuth and sun zenith (in radians).
Or a RasterStack/Brick with pre-calculated slope and aspect (see \link[raster]{terrain}). In this case the layers must be named 'slope' and 'aspect'.}

\item{method}{Character. One of c("cos", "avgcos", "minnaert", "C", "stat", "cosi").}

\item{stratImg}{RasterLayer by which to stratify by, e.g. NDVI. Defaults to 'slope'}

\item{stratMeth}{Character. One of c("noStrat", "stratEqualBins", "stratQuantiles").}

\item{cosi}{Raster*. Optional pre-calculated ilumination map. Run topCor with method="cosi" to calculate one.}

\item{nStrat}{Integer. Number of bins or quantiles to stratify by. If a bin has less than 50 samples it will be merged with the next bin.}
}
\description{
Topographic Illumination Correction
}

